[core] Prevent unstable chunks in size snapshot
