+++ scripts/sizeSnapshot/webpack.config.js
@@ -1,126 +1,14 @@
-const globCallback = require('glob');
 const path = require('path');
 const CompressionPlugin = require('compression-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
-const { promisify } = require('util');
-
-const glob = promisify(globCallback);
 
 const workspaceRoot = path.join(__dirname, '..', '..');
 
-async function getSizeLimitBundles() {
-  const corePackagePath = path.join(workspaceRoot, 'packages/material-ui/build');
-  const coreComponents = (await glob(path.join(corePackagePath, '[A-Z]*/index.js'))).map(
-    (componentPath) => {
-      const componentName = path.basename(path.dirname(componentPath));
-      let entryName = componentName;
-      // adjust for legacy names
-      if (componentName === 'Paper') {
-        entryName = '@material-ui/core/Paper.esm';
-      } else if (componentName === 'TextareaAutosize') {
-        entryName = '@material-ui/core/Textarea';
-      } else if (['Popper'].indexOf(componentName) !== -1) {
-        entryName = `@material-ui/core/${componentName}`;
-      }
-
-      return {
-        name: entryName,
-        webpack: true,
-        path: path.relative(workspaceRoot, path.dirname(componentPath)),
-      };
-    },
-  );
-
-  const labPackagePath = path.join(workspaceRoot, 'packages/material-ui-lab/build');
-  const labComponents = (await glob(path.join(labPackagePath, '[A-Z]*/index.js'))).map(
-    (componentPath) => {
-      const componentName = path.basename(path.dirname(componentPath));
-
-      return {
-        name: componentName,
-        webpack: true,
-        path: path.relative(workspaceRoot, path.dirname(componentPath)),
-      };
-    },
-  );
-
-  return [
-    {
-      name: '@material-ui/core',
-      webpack: true,
-      path: path.join(path.relative(workspaceRoot, corePackagePath), 'index.js'),
-    },
-    {
-      name: '@material-ui/lab',
-      webpack: true,
-      path: path.join(path.relative(workspaceRoot, labPackagePath), 'index.js'),
-    },
-    {
-      name: '@material-ui/styles',
-      webpack: true,
-      path: 'packages/material-ui-styles/build/index.js',
-    },
-    {
-      name: '@material-ui/system',
-      webpack: true,
-      path: 'packages/material-ui-system/build/esm/index.js',
-    },
-    ...coreComponents,
-    {
-      name: '@material-ui/core/styles/createMuiTheme',
-      webpack: true,
-      path: 'packages/material-ui/build/styles/createMuiTheme.js',
-    },
-    {
-      name: 'colorManipulator',
-      webpack: true,
-      path: 'packages/material-ui/build/styles/colorManipulator.js',
-    },
-    ...labComponents,
-    {
-      name: 'useAutocomplete',
-      webpack: true,
-      path: 'packages/material-ui-lab/build/useAutocomplete/index.js',
-    },
-    {
-      name: '@material-ui/core/useMediaQuery',
-      webpack: true,
-      path: 'packages/material-ui/build/useMediaQuery/index.js',
-    },
-    {
-      name: '@material-ui/core/useScrollTrigger',
-      webpack: true,
-      path: 'packages/material-ui/build/useScrollTrigger/index.js',
-    },
-    {
-      name: '@material-ui/utils',
-      webpack: true,
-      path: 'packages/material-ui-utils/build/esm/index.js',
-    },
-    // TODO: Requires webpack v5
-    // Resolution of webpack/acorn to 7.x is blocked by nextjs (https://github.com/vercel/next.js/issues/11947)
-    // {
-    //   name: '@material-ui/core.modern',
-    //   webpack: true,
-    //   path: path.join(path.relative(workspaceRoot, corePackagePath), 'modern/index.js'),
-    // },
-    {
-      name: '@material-ui/core.legacy',
-      webpack: true,
-      path: path.join(path.relative(workspaceRoot, corePackagePath), 'legacy/index.js'),
-    },
-  ];
-}
-
 module.exports = async function webpackConfig() {
-  const entries = await getSizeLimitBundles();
-  const entry = entries.reduce((acc, bundle) => {
-    acc[bundle.name] = path.join(workspaceRoot, bundle.path);
-    return acc;
-  }, {});
-
   const config = {
-    entry,
+    cache: {
+      type: 'memory',
+    },
     // ideally this would be computed from the bundles peer dependencies
     externals: /^(react|react-dom|react\/jsx-runtime)$/,
     mode: 'production',
